\name{colPercents}
\alias{colPercents}
\alias{rowPercents}

\title{Row and Column Percentage Tables}
\description{
  Percentage a matrix or higher-dimensional array of frequency counts by
  rows or columns.
}
\usage{
colPercents(tab)
rowPercents(tab)
}

\arguments{
  \item{tab}{a matrix or higher-dimensional array of frequency counts.}
}

}
\value{
  Returns an array of the same size and shape as \code{tab} percentaged by
  rows or columns.
}

\author{John Fox \email{jfox@mcmaster.ca}}


\keyword{misc}

