% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-menu.R
\name{Compute}
\alias{Compute}
\title{Rcmdr Compute Dialog}
\description{
Rcmdr Compute Dialog
}
\details{
The compute dialog is used to compute new variables.

The name of the new variable must be a valid R object name (consisting only of upper and lower-case letters, numerals, and periods, and not starting with a numeral).

Enter an R expression in the box at the right.
The expression is evaluated using the active data set.
You can double-click in the variable-list box to enter variable names in the expression.
The expression must evaluate to a valid variable, which is added to the active data set.
}
\seealso{
\code{\link[base]{Arithmetic}}
}
\author{
John Fox
}
\keyword{manip}
