% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-menu.R
\name{dropUnusedFactorLevels}
\alias{dropUnusedFactorLevels}
\title{Drop Unused Factor Levels}
\description{
Drop Unused Factor Levels
}
\details{
Following filtering or recoding procedures, certain factor levels may exhibit zero frequencies. This menu option allows users to remove such unobserved levels from the factor variable.

This menu option calls the \code{\link[base:droplevels]{droplevles}} function from the \pkg{base} package.
}
\keyword{manip}
