% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-models-menu.R
\name{linearRegressionModel}
\alias{linearRegressionModel}
\title{Rcmdr Linear Regression Model}
\description{
Rcmdr Linear Regression Model
}
\details{
This menu option can be used to build a linear regression model. This is a linear model in which the response and all explanatory variables are numeric.

After selecting this option, in the new dialog, optionally, change the model name in the first field.

Select a response variable from the left list box, followed by one or more variables as explanatory variables from the right list box.

You may also specify a list of indices/row names to exclude in the third field.

Or a subset of the data by providing a subset expression, in fourth field.

Clicking the OK button produces a call to the \code{\link[stats]{lm}} function in the \pkg{stats} package to build the regression model, followed by a call to summarize and print the results.
}
\seealso{
See \code{\link{linearModel}} for a more detailed description of this menu option.
}
\keyword{models}
