% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-menu.R
\name{Options}
\alias{Options}
\title{R Commander Options}
\usage{
Options()
}
\description{
R Commander Options
}
\details{
This dialog allows you to change the options of the running instance of R Commander.

The current Rcmdr options can be saved via \emph{Tools -> Save Rcmdr options...} menu.
Changes can be made to the \code{.Rprofile} file before saving the options.

See \code{\link{Startup}} for a description of the \file{.Rprofile} file and the R startup process in general.

See \code{\link{Commander}} for a description of the various R Commander options.
}
\author{
Manuel Munoz-Marquez
}
\keyword{misc}
