\name{Rcmdr-internal}
\alias{Rcmdr-internal}
\alias{tcl}
\alias{tclRequire}
\alias{tclServiceMode}
\alias{tclVar}
\alias{tclvalue<-}
\alias{tk_select.list}
\alias{tkbind}
\alias{tkcget}
\alias{tkcheckbutton}
\alias{tkchooseDirectory}
\alias{tkconfigure}
\alias{tkcurselection}
\alias{tkdelete}
\alias{tkdestroy}
\alias{tkentry}
\alias{tkframe}
\alias{tkgetOpenFile}
\alias{tkgetSaveFile}
\alias{tkgrab.release}
\alias{tkgrab.set}
\alias{tkgrid}
\alias{tkgrid.columnconfigure}
\alias{tkgrid.configure}
\alias{tkgrid.rowconfigure}
\alias{tkicursor}
\alias{tkindex}
\alias{tkinsert}
\alias{tkinvoke}
\alias{tkitemconfigure}
\alias{tklabel}
\alias{tklistbox}
\alias{tkmark.set}
\alias{tkmessageBox}
\alias{tkpack}
\alias{tkradiobutton}
\alias{tkraise}
\alias{tkscale}
\alias{tkscrollbar}
\alias{tkselection.clear}
\alias{tkselection.present}
\alias{tkselection.set}
\alias{tkset}
\alias{tktag.configure}
\alias{tktext}
\alias{tktoplevel}
\alias{tkwait.window}
\alias{tkwm.deiconify}
\alias{tkwm.geometry}
\alias{tkwm.iconbitmap}
\alias{tkwm.protocol}
\alias{tkwm.resizable}
\alias{tkwm.title}
\alias{tkwm.transient}
\alias{tkxview}
\alias{tkxview.moveto}
\alias{tkyview}
\alias{tkyview.scroll}
\alias{ttklabelframe}
\alias{ttkcheckbutton}

\title{
\pkg{tcltk} Functions Re-exported by the \pkg{Rcmdr} Package.
}
\description{
For technical reasons involving \pkg{Rcmdr} plug-in packages, the \pkg{Rcmdr} package
re-exports a number of functions from the \pkg{tcltk} package that are used by \pkg{Rcmdr}
utilities.
}

\author{John Fox \email{jfox@mcmaster.ca}}

\seealso{
 \code{\link{tcltk}}
}

\keyword{misc}
