% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RcmdrPager}
\alias{RcmdrPager}
\title{Pager for Text Files}
\usage{
RcmdrPager(file, header, title, delete.file)
}
\arguments{
\item{file}{character vector of file(s) to be displayed.}

\item{header}{for the beginning of each file.}

\item{title}{for window.}

\item{delete.file}{delete file(s) on close.}
}
\description{
Pager for Text Files
}
\details{
This is a slightly modified version of the tkpager, changed to use the Rcmdr monospaced font and a white background.
}
\seealso{
\code{\link[tcltk]{tkpager}}.
}
\keyword{misc}
