% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-menu.R
\name{readDataFromPackage}
\alias{readDataFromPackage}
\title{Read data from package}
\description{
Read data from package
}
\details{
This menu option can be used to load data sets provided by currently attached packages.

Selecting the menu items \code{Data} -> \code{Data in packages} -> \code{Read data set from an attached package...} opens a new dialog window.

Double-clicking an item in the left list box displays the available data sets for the selected package in the right list box.
Once a data set is selected, clicking \code{OK} loads and activates it.

Additionally, documentation for a selected data set can be accessed by clicking the \code{Help on selected data set} button.

If the name of the data set is already known, it can be entered directly into the "Enter name of data set:" field.
Note that this field is only available before a package has been selected from the list.

For example, double-clicking \code{carData} in the left list, selecting \code{Chile} in the right list, and clicking \code{OK} will load and activate the \code{Chile} data set from the \pkg{carData} package.

This menu option calls the \code{\link[utils]{data}} function from the \code{\link[utils]{utils-package}} package.
}
