% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcmdr-internal.R, R/Rcmdr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{showData}
\alias{.Tcl}
\alias{tcl}
\alias{tkfocus}
\alias{ttkcheckbutton}
\alias{tclRequire}
\alias{tclServiceMode}
\alias{tclvalue<-}
\alias{tclVar}
\alias{tkbind}
\alias{tkcget}
\alias{tkcheckbutton}
\alias{tkconfigure}
\alias{tkchooseDirectory}
\alias{tkcurselection}
\alias{tkdelete}
\alias{tkdestroy}
\alias{tkentry}
\alias{tkframe}
\alias{tkgrab.release}
\alias{tkgrab.set}
\alias{tkgetOpenFile}
\alias{tkgetSaveFile}
\alias{tkgrid}
\alias{tkgrid.columnconfigure}
\alias{tkgrid.configure}
\alias{tkgrid.rowconfigure}
\alias{tkicursor}
\alias{tkindex}
\alias{tkinsert}
\alias{tkinvoke}
\alias{tkitemconfigure}
\alias{tklabel}
\alias{tklistbox}
\alias{tkmark.set}
\alias{tkmessageBox}
\alias{tkpack}
\alias{tkradiobutton}
\alias{tkraise}
\alias{tkscale}
\alias{tkscrollbar}
\alias{tk_select.list}
\alias{tkselection.clear}
\alias{tkselection.present}
\alias{tkselection.set}
\alias{tkset}
\alias{tktag.configure}
\alias{tktext}
\alias{tktoplevel}
\alias{tkwait.window}
\alias{tkwm.deiconify}
\alias{tkwm.iconbitmap}
\alias{tkwm.geometry}
\alias{tkwm.protocol}
\alias{tkwm.resizable}
\alias{tkwm.title}
\alias{tkwm.transient}
\alias{tkxview.moveto}
\alias{tkxview}
\alias{tkyview}
\alias{tkyview.scroll}
\alias{ttklabelframe}
\title{\pkg{tcltk} Functions Re-exported by the \pkg{Rcmdr} Package.}
\details{
For technical reasons involving \pkg{Rcmdr} plug-in packages, the \pkg{Rcmdr} package re-exports a number of functions from the \pkg{tcltk} package that are used by \pkg{Rcmdr} utilities.

It also re-exports the \code{showdata} function from the \pkg{relimp} package, which is uses internally to display data sets.
}
\seealso{
\code{\link{tcltk}}.
}
\author{
John Fox
}
\keyword{internal}
\keyword{misc}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{relimp}{\code{\link[relimp]{showData}}}

  \item{tcltk}{\code{\link[tcltk:TclInterface]{.Tcl}}, \code{\link[tcltk:TkCommands]{tcl}}, \code{\link[tcltk:TclInterface]{tclRequire}}, \code{\link[tcltk]{tclServiceMode}}, \code{\link[tcltk:TclInterface]{tclvalue<-}}, \code{\link[tcltk:TclInterface]{tclVar}}, \code{\link[tcltk]{tk_select.list}}, \code{\link[tcltk:TkCommands]{tkbind}}, \code{\link[tcltk:TkWidgetcmds]{tkcget}}, \code{\link[tcltk:TkWidgets]{tkcheckbutton}}, \code{\link[tcltk:TkCommands]{tkchooseDirectory}}, \code{\link[tcltk:TkWidgetcmds]{tkconfigure}}, \code{\link[tcltk:TkWidgetcmds]{tkcurselection}}, \code{\link[tcltk:TkWidgetcmds]{tkdelete}}, \code{\link[tcltk:TclInterface]{tkdestroy}}, \code{\link[tcltk:TkWidgets]{tkentry}}, \code{\link[tcltk:TkCommands]{tkfocus}}, \code{\link[tcltk:TkWidgets]{tkframe}}, \code{\link[tcltk:TkCommands]{tkgetOpenFile}}, \code{\link[tcltk:TkCommands]{tkgetSaveFile}}, \code{\link[tcltk:TkCommands]{tkgrab.release}}, \code{\link[tcltk:TkCommands]{tkgrab.set}}, \code{\link[tcltk:TkCommands]{tkgrid}}, \code{\link[tcltk:TkCommands]{tkgrid.columnconfigure}}, \code{\link[tcltk:TkCommands]{tkgrid.configure}}, \code{\link[tcltk:TkCommands]{tkgrid.rowconfigure}}, \code{\link[tcltk:TkWidgetcmds]{tkicursor}}, \code{\link[tcltk:TkWidgetcmds]{tkindex}}, \code{\link[tcltk:TkWidgetcmds]{tkinsert}}, \code{\link[tcltk:TkWidgetcmds]{tkinvoke}}, \code{\link[tcltk:TkWidgetcmds]{tkitemconfigure}}, \code{\link[tcltk:TkWidgets]{tklabel}}, \code{\link[tcltk:TkWidgets]{tklistbox}}, \code{\link[tcltk:TkWidgetcmds]{tkmark.set}}, \code{\link[tcltk:TkCommands]{tkmessageBox}}, \code{\link[tcltk:TkCommands]{tkpack}}, \code{\link[tcltk:TkWidgets]{tkradiobutton}}, \code{\link[tcltk:TkCommands]{tkraise}}, \code{\link[tcltk:TkWidgets]{tkscale}}, \code{\link[tcltk:TkWidgets]{tkscrollbar}}, \code{\link[tcltk:TkWidgetcmds]{tkselection.clear}}, \code{\link[tcltk:TkWidgetcmds]{tkselection.present}}, \code{\link[tcltk:TkWidgetcmds]{tkselection.set}}, \code{\link[tcltk:TkWidgetcmds]{tkset}}, \code{\link[tcltk:TkWidgetcmds]{tktag.configure}}, \code{\link[tcltk:TkWidgets]{tktext}}, \code{\link[tcltk:TkWidgets]{tktoplevel}}, \code{\link[tcltk:TkCommands]{tkwait.window}}, \code{\link[tcltk:TkCommands]{tkwm.deiconify}}, \code{\link[tcltk:TkCommands]{tkwm.geometry}}, \code{\link[tcltk:TkCommands]{tkwm.iconbitmap}}, \code{\link[tcltk:TkCommands]{tkwm.protocol}}, \code{\link[tcltk:TkCommands]{tkwm.resizable}}, \code{\link[tcltk:TkCommands]{tkwm.title}}, \code{\link[tcltk:TkCommands]{tkwm.transient}}, \code{\link[tcltk:TkWidgetcmds]{tkxview}}, \code{\link[tcltk:TkWidgetcmds]{tkxview.moveto}}, \code{\link[tcltk:TkWidgetcmds]{tkyview}}, \code{\link[tcltk:TkWidgetcmds]{tkyview.scroll}}, \code{\link[tcltk:TkWidgets]{ttkcheckbutton}}, \code{\link[tcltk:TkWidgets]{ttklabelframe}}}
}}

