% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-menu.R
\name{SelectRows}
\alias{SelectRows}
\title{Select Rows from Active Data Set}
\description{
Select Rows from Active Data Set
}
\details{
This menu option can be used to build a new data set by selecting rows by name.

In the "Indices or names of row(s) to select" field, enter the names of the rows you wish to keep, separated by spaces.
If a row name contains a space, it must be enclosed in single or double quotes.
In all other cases, quotes are not required.
You can also specify a name for the resulting data set in the "Name for new data set" field.
The new data set will automatically become the active data set.

If the data set has no associated row names, the original row indices are used as names.
Keep in mind that subsequent operations do not alter these row names; therefore, the numbers will always refer to the original row positions.

This menu option produces a call to the \code{\link[base]{[.data.frame}} method from the \pkg{base} package.
}
\seealso{
See also \code{\link{RemoveRows}}.
}
\keyword{manip}
