% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseRegression.R
\name{stepwiseRegression}
\alias{stepwiseRegression}
\title{Stepwise Model Selection}
\description{
Stepwise Model Selection
}
\details{
This menu option performs stepwise model selection using either BIC or AIC for active model.

The selection direction determines whether the process starts with the full set of variables and removes them (backward step) or starts with an empty model and adds variables to it (forward step).
If \code{"backward/forward"} or \code{"forward/backward"} is selected, the backward and forward steps are alternated.

The selection criterion can be \code{"BIC"} (Bayesian Information Criterion) or \code{"AIC"} (Akaike Information Criterion).

Note that \code{\link[RcmdrMisc]{stepwise}} labels the criterion in the output as \code{"AIC"} regardless of which criterion is actually employed.

This menu option calls the \code{\link[RcmdrMisc]{stepwise}} function in the \pkg{RcmdrMisc} package.
}
\seealso{
See also \code{\link{lm}}, \code{\link{linearModel}}.
}
\keyword{models}
