\name{Rcmdr-package}
\alias{Rcmdr-package}
\alias{Rcmdr}
\docType{package}
\title{
R Commander
}
\description{
A platform-independent basic-statistics GUI (graphical
user interface) for R, based on the tcltk package.
}
\details{
\tabular{ll}{
Package: \tab Rcmdr\cr
Version: \tab 1.1-6\cr
Date: \tab 2006/2/5\cr
Depends: \tab R (>= 2.1.0), tcltk\cr
Suggests: \tab abind, car (>= 1.0-15), effects (>= 1.0-7), foreign, grid,
lattice, lmtest, MASS, mgcv, multcomp, nlme, nnet, relimp\cr
LazyLoad: \tab no\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.r-project.org,
http://socserv.socsci.mcmaster.ca/jfox/Misc/Rcmdr/\cr
}

Index:
\preformatted{
Commander               R Commander
Compute                 Rcmdr Compute Dialog
Confint                 Confidence Intervals for Model Coefficients
Hist                    Plot a Histogram
KMeans                  K-Means Clustering Using Multiple Random Seeds
Rcmdr.Utilities         Rcmdr Utility Functions
Rcmdr.sciviews-specific
                        Rcmdr SciViews-specific Functions
RcmdrPager              Pager for Text Files
Recode                  Rcmdr Recode Dialog
Scatter3DDialog         Rcmdr 3D Scatterplot Dialog
aboutRcmdr              About the Rcmdr Package
assignCluster           Append a Cluster Membership Variable to a
                        Dataframe
bin.var                 Bin a Numeric Varisible
colPercents             Row and Column Percentage Tables
generalizedLinearModel
                        Rcmdr Generalized Linear Model Dialog
hierarchicalCluster     Rcmdr Hierarchical Clustering Dialog
linearModel             Rcmdr Linear Model Dialog
partial.cor             Partial Correlations
plotMeans               Plot Means for One or Two-Way Layout
reliability             Reliability of a Composite Scale
scatter3d               Three-Dimensional Scatterplots and Point
                        Identification
stem.leaf               Stem-and-Leaf Display
}
}
\author{
John Fox <jfox@mcmaster.ca>, with contributions from Michael
Ash, Philippe Grosjean, Martin Maechler, Dan Putler, and Peter
Wolf.

Maintainer: John Fox <jfox@mcmaster.ca>
}
\keyword{ package }
