\name{Rcmdr-package}
\alias{Rcmdr-package}
\alias{Rcmdr}
\docType{package}
\title{
R Commander
}
\description{
A platform-independent basic-statistics GUI (graphical
user interface) for R, based on the tcltk package.
}

\references{
Fox, J. (2017)
\emph{Using the R Commander: A Point-and-Click Interface for R}.
Chapman and Hall/CRC Press.

Fox, J. (2005) 
The R Commander: A Basic Statistics Graphical User Interface to R. 
\emph{Journal of Statistical Software}, \bold{14(9)}: 1--42.
}

\section{Translations}{
The R Commander comes with translations from English into several other languages. I am grateful to the following individuals and groups for preparing these translations:
Basque, Jose Ramon Rueda;
Brazilian Portuguese, Adriano Azevedo-Filho and Marilia Sa Carvalho;
Catalan, Manel Salamero;
Chinese, Tsungwu Ho, Frank C. S. Liu, and Cheng-shun Lee;
Chinese (Simplified), Shulin Yang;
French, Philippe Grosjean and Milan Bouchet-Valat;
Galician, Anton Meixome;
German: Friedrich Leisch and Gerhard Schoen;
Greek: Vasileios Dimitropoulos, Anastasios Vikatos, and Andreas Vikatos;
Indonesian, I Made Tirta;
Italian, Stefano Calza;
Japanese, Takaharu Araki;
Korean, Chel Hee Lee, Dae-Heung Jang, and Shin Jong-Hwa;
Polish, Lukasz Daniel;
Romanian, Adrian Dusa;
Russian, Alexey Shipunov;
Slovenian, Jaro Lajovic and Matjaz Jeran;
Spanish, Spanish R-UCA Project, http://knuth.uca.es/R.

Note: Many of these translations are not up-to-date. If you encounter an out-of-date translation and wish to take over updating and maintaining it, please contract the R Commander developers.

}

\author{
John Fox <jfox@mcmaster.ca>, Manuel Munoz Marquez, and Milan Bouchet-Valat, with contributions from Liviu Andronic, Michael Ash, Theophilius Boye, Stefano Calza, Andy Chang, Vilmantas Gegzna,
Philippe Grosjean, Richard Heiberger, G. Jay Kerns, Renaud Lancelot, Matthieu Lesnoff, Uwe Ligges, Samir Messad, Martin Maechler, Robert Muenchen, Duncan Murdoch, Erich Neuwirth, Dan Putler, Brian Ripley, Miroslav Ristic,  Peter Wolf, and Kevin Wright.

Maintainer: John Fox <jfox@mcmaster.ca>
}
\keyword{ package }
