\name{Rcmdr-package}
\alias{Rcmdr-package}
\alias{Rcmdr}
\docType{package}
\title{
R Commander
}
\description{
A platform-independent basic-statistics GUI (graphical
user interface) for R, based on the tcltk package.
}
\details{
\tabular{ll}{
Package: \tab Rcmdr\cr
Version: \tab 2.0-1\cr
Date: \tab 2013/11/08\cr
Depends: \tab  R (>= 3.0.0), grDevices, utils, splines\cr
Imports: \tab tcltk\cr
Suggests: \tab abind, aplpack, car (>= 2.0-18), colorspace, effects (>= 1.0-7),
    e1071, foreign, grid, Hmisc, knitr, lattice, leaps, lmtest, MASS, markdown, mgcv, 
    multcomp (>= 0.991-2), nlme, nnet, relimp, rgl, RODBC, sem (>= 2.1-1)\cr
ByteCompile: \tab yes\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org,
http://socserv.socsci.mcmaster.ca/jfox/Misc/Rcmdr/\cr
}
}

\references{
Fox, J. (2005) 
The R Commander: A Basic Statistics Graphical User Interface to R. 
\emph{Journal of Statistical Software}, \bold{14(9)}: 1--42.

Fox, J. (2007)
Extending the R Commander by "plug in" packages. 
\emph{R News}, \bold{7(3)}: 46--52.
}

\section{Translations}{
The R Commander comes with translations from English into several other languages. I am grateful to the following individuals and groups for preparing these translations:
Brazilian Portuguese, Adriano Azevedo-Filho and Marilia Sa Carvalho;
Catalan, Manel Salamero;
Chinese, Frank C. S. Liu and Cheng-shun Lee;
French, Philippe Grosjean and Milan Bouchet-Valat;
German: Gerhard Schoen;
Indonesian, I Made Tirta;
Italian, Stefano Calza;
Japanese, Takaharu Araki;
Korean, Chel Hee Lee, Dae-Heung Jang, and Shin Jong-Hwa;
Polish, Marcin Kozak;
Romanian, Adrian Dusa;
Russian, Alexey Shipunov;
Slovenian, Jaro Lajovic and Matjaz Jeran;
Spanish, Spanish R-UCA Project, http://knuth.uca.es/R.
}

\author{
John Fox <jfox@mcmaster.ca> and Milan Bouchet-Valat, with contributions from Liviu Andronic, Michael Ash, Theophilius Boye, Stefano Calza, Andy Chang, Philippe Grosjean, Richard Heiberger, G. Jay Kerns, Renaud Lancelot, Matthieu Lesnoff, Uwe Ligges, Samir Messad, Martin Maechler, Robert Muenchen, Duncan Murdoch, Erich Neuwirth, Dan Putler, Brian Ripley, Miroslav Ristic,  Peter Wolf, and Kevin Wright.

Maintainer: John Fox <jfox@mcmaster.ca>
}
\keyword{ package }
